+++ lib/ratehawk/index.ts
@@ -1,4 +1,5 @@
 export * as hotel from "./services/hotel";
 export { ratehawkRequest, RatehawkError } from "./client";
 export { config as ratehawkConfig } from "./config";
+export { rhImage } from "./image";
 export type * from "./types";